// src/interfaces/IResponseData.ts
export interface IResponseData {
  timestamp: string;
  status: boolean;
  statusCode: number;
  method: string;
  path: string;
  query: Record<string, any>;
  headers: Record<string, any>;
  body: any;
  message: string;
  suggestion: string;
}

export interface IGetHealthResponseData {
  timestamp: string;
  status: boolean;
  statusCode: number;
  method: string;
  path: string;
  query: Record<string, any>;
  headers: Record<string, any>;
  body: any;
  data: {
    cpuUsage: {
      name: string;
      value: string;
    };
    memoryUsage: {
      name: string;
      value: string;
    };
  };
}

export interface ICreateServiceOrderX1ResponseData {
  timestamp: string;
  status: boolean;
  statusCode: number;
  method: string;
  path: string;
  query: Record<string, any>;
  headers: Record<string, any>;
  body: any;
  data: unknown;
}